requirements.txt
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.10.5
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-utils-six==2.0
djangorestframework==3.15.2
pillow==11.1.0
pycodestyle==2.12.1
python-dotenv==1.1.1
sentry-sdk==2.40.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
